removed draggable list
