+++ src/templates/components/organisms/Panels/PanelMobileComponent.vue
@@ -16,13 +16,11 @@
 <script>
 import { defineComponent, toRef, computed, onMounted, ref } from "vue";
 import { useStore } from "vuex";
-import DraggableList from "@/templates/components/molecules/DraggableList/DraggableList.vue";
 import PanelResults from "@/templates/components/organisms/Panels/PanelResults.vue";
 
 export default defineComponent({
   name: "PanelMobileComponent",
   components: {
-    DraggableList,
     PanelResults,
   },
   props: {
